Add Visual Studio temp files to gitignore
There's been a few instances where this project has been opened in Visual Studio. When that happens Visual Studio creates a `.vs/` directory with temporary files about the project which gets included when those users contribute their changes.

This adds the `.vs/` directory to the `.gitignore` file so that others can avoid contributing that directory.